refactor: auth 패키지의 dto를 내부 dto로 수정
