--- backend/src/main/java/codesquad/team01/issuetracker/auth/dto/GitHubEmail.java
@@ -1,9 +1,0 @@
-package codesquad.team01.issuetracker.auth.dto;
-
-public record GitHubEmail(
-	String email,
-	boolean primary,
-	boolean verified,
-	String visibility
-) {
-}
