--- backend/src/main/java/codesquad/team01/issuetracker/auth/dto/GitHubUser.java
@@ -1,10 +1,0 @@
-package codesquad.team01.issuetracker.auth.dto;
-
-// 깃헙이 주는 사용자의 정보를 담는 dto
-public record GitHubUser(
-	long id,
-	String githubId,
-	String avatarUrl,
-	String email
-) {
-}
