--- backend/src/main/java/codesquad/team01/issuetracker/auth/dto/LoginRequest.java
@@ -1,4 +1,0 @@
-package codesquad.team01.issuetracker.auth.dto;
-
-public class LoginRequest {
-}
